+++ geyser-plugin-manager/src/geyser_plugin_service.rs
@@ -109,8 +109,10 @@ impl GeyserPluginService {
             .load()
             .deshred_transaction_notifications_enabled()
             || geyser_plugin_always_enabled;
-        let entry_notifications_enabled =
-            plugin_manager.load().entry_notifications_enabled() || geyser_plugin_always_enabled;
+        let entry_or_block_footer_notifications_enabled =
+            plugin_manager.load().entry_notifications_enabled()
+                || plugin_manager.load().block_footer_notifications_enabled()
+                || geyser_plugin_always_enabled;
 
         let accounts_update_notifier: Option<AccountsUpdateNotifier> =
             if account_data_notifications_enabled {
@@ -140,12 +142,13 @@ impl GeyserPluginService {
                 None
             };
 
-        let entry_notifier: Option<EntryNotifierArc> = if entry_notifications_enabled {
-            let entry_notifier = EntryNotifierImpl::new(plugin_manager.clone());
-            Some(Arc::new(entry_notifier))
-        } else {
-            None
-        };
+        let entry_notifier: Option<EntryNotifierArc> =
+            if entry_or_block_footer_notifications_enabled {
+                let entry_notifier = EntryNotifierImpl::new(plugin_manager.clone());
+                Some(Arc::new(entry_notifier))
+            } else {
+                None
+            };
 
         let (slot_status_observer, block_metadata_notifier, slot_status_notifier): (
             Option<SlotStatusObserver>,
@@ -154,7 +157,7 @@ impl GeyserPluginService {
         ) = if account_data_notifications_enabled
             || transaction_notifications_enabled
             || deshred_transaction_notifications_enabled
-            || entry_notifications_enabled
+            || entry_or_block_footer_notifications_enabled
         {
             let slot_status_notifier = SlotStatusNotifierImpl::new(plugin_manager.clone());
             let slot_status_notifier = Arc::new(RwLock::new(slot_status_notifier));